# recipes/service/spec/dist/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "recipes API",
    "version": "1.0.0",
    "description": "API specification for the recipes service"
  },
  "paths": {
    "/recipes": {
      "get": {
        "summary": "List recipes. Either (a) collectionId: list recipes in that collection (auth required). Or (b) publicOnly=true: all public recipes across collections (no auth).",
        "operationId": "listRecipes",
        "tags": [
          "no-auth"
        ],
        "parameters": [
          {
            "name": "collectionId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Id of the collection to list recipes from. Omit when publicOnly=true."
          },
          {
            "name": "publicOnly",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            },
            "description": "When true, return all public recipes across all collections. Used by root (logged-out) app. Do not send collectionId when using this."
          }
        ],
        "responses": {
          "200": {
            "description": "List of recipes (public only when not admin, includes private for admins).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/recipe"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - provide either collectionId or publicOnly=true, not both nor neither.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - collection listing requires authentication.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - not a member of the collection or insufficient role.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create recipe (and optionally initial version). Admin only.",
        "operationId": "createRecipe",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "collectionId",
                  "title",
                  "subtitle",
                  "isPublic"
                ],
                "properties": {
                  "collectionId": {
                    "type": "string",
                    "description": "Id of the collection this recipe belongs to"
                  },
                  "title": {
                    "type": "string",
                    "description": "Recipe title"
                  },
                  "subtitle": {
                    "type": "string",
                    "description": "Subtitle shown on menus and list views"
                  },
                  "description": {
                    "type": "string",
                    "nullable": true,
                    "description": "Optional longer description shown on recipe detail"
                  },
                  "isPublic": {
                    "type": "boolean",
                    "description": "Whether the recipe is visible to non-admin users"
                  },
                  "initialVersion": {
                    "type": "object",
                    "description": "Optional initial version content for the first version",
                    "properties": {
                      "content": {
                        "$ref": "#/components/schemas/content"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Created recipe and initial RecipeVersion if provided.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "recipe"
                  ],
                  "properties": {
                    "recipe": {
                      "$ref": "#/components/schemas/recipe"
                    },
                    "initialVersion": {
                      "$ref": "#/components/schemas/recipe-version",
                      "description": "Present when initialVersion was supplied in the request."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - collectionId is required in request body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - missing or invalid auth headers, or not logged in.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - user does not have required privileges (admin only).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        }
      }
    },
...
    "/recipes/{id}/files/from-unsplash": {
      "post": {
        "summary": "Add a recipe file from an Unsplash photo. Backend fetches photo, tracks download, stores file and Unsplash user; returns RecipeFileInfo with unsplashAttribution when applicable.",
        "operationId": "filesFromUnsplashRecipes",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Recipe id"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/add-recipe-file-from-unsplash-request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Created recipe file metadata (includes unsplashAttribution when from Unsplash).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/recipe-file-info"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - missing or invalid auth headers, or not logged in.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - user does not have required privileges (admin only).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "404": {
            "description": "Not Found - recipe not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        }
      }
    },
...
    "/collections/{id}/members/{memberId}": {
      "put": {
        "summary": "Update a member's role. Owner only. Cannot demote the creator (permanent owner).",
        "operationId": "membersUpdateCollections",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Collection id"
          },
          {
            "name": "memberId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Member record id"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "role"
                ],
                "properties": {
                  "role": {
                    "type": "string",
                    "enum": [
                      "owner",
                      "editor",
                      "viewer"
                    ],
                    "description": "New role for the member (creator's role cannot be changed from owner)",
                    "example": "viewer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated member.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "member"
                  ],
                  "properties": {
                    "member": {
                      "$ref": "#/components/schemas/collection-member",
                      "description": "The updated collection member."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - cannot demote the creator (permanent owner).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - missing or invalid auth headers, or not logged in.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - caller must be an owner of the collection.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "404": {
            "description": "Not Found - collection or member does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Remove a member from a collection. Owner only. Cannot remove the creator (permanent owner); returns 400 if attempted.",
        "operationId": "membersRemoveCollections",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Collection id"
          },
          {
            "name": "memberId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Member record id"
          }
        ],
        "responses": {
          "204": {
            "description": "Member removed."
          },
          "400": {
            "description": "Bad Request - cannot remove the creator (permanent owner).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - missing or invalid auth headers, or not logged in.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - caller must be an owner of the collection.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "404": {
            "description": "Not Found - collection or member does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "recipe": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the recipe",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "title": {
            "type": "string",
            "description": "Recipe title",
            "example": "Classic Chocolate Chip Cookies"
          },
          "subtitle": {
            "type": "string",
            "description": "Subtitle shown on menus and list views",
            "example": "Crispy edges, chewy centers"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "description": "Optional longer description shown on recipe detail (not on menu)",
            "example": "A crowd-pleasing recipe that works every time. Best with room-temperature butter."
          },
          "isPublic": {
            "type": "boolean",
            "description": "Whether the recipe is visible to non-admin users",
            "example": true
          },
          "createdBy": {
            "type": "string",
            "description": "User id of the creator",
            "example": "a1b2c3d4-e89b-12d3-a456-426614174001"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "When the recipe was created",
            "example": "2023-01-15T14:30:00Z"
          },
          "updatedBy": {
            "type": "string",
            "description": "User id of the last updater",
            "example": "a1b2c3d4-e89b-12d3-a456-426614174001"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "When the recipe was last updated",
            "example": "2023-02-01T09:00:00Z"
          },
          "currentVersionId": {
            "type": "string",
            "description": "Optional id of the current/latest version for display",
            "example": "b2c3d4e5-e89b-12d3-a456-426614174002"
          }
        },
        "required": [
          "id",
          "title",
          "subtitle",
          "isPublic",
          "createdBy",
          "createdAt",
          "updatedBy",
          "updatedAt"
        ]
      },
      "error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "A short, machine-readable error code, for when HTTP status codes are not sufficient."
          },
          "message": {
            "type": "string",
            "description": "A human-readable description of the error.",
            "example": "The requested resource could not be found."
          }
        }
      },
      "content": {
        "type": "object",
        "description": "Version body — structured ingredients and markdown instructions",
        "properties": {
          "ingredients": {
            "type": "array",
            "description": "List of ingredients; quantity and unit are strings (e.g. \"2+1/3\", \"cups\")",
            "items": {
              "type": "object",
              "required": [
                "name",
                "quantity",
                "unit"
              ],
              "properties": {
                "name": {
                  "type": "string",
                  "description": "Ingredient name",
                  "example": "All-purpose flour"
                },
                "quantity": {
                  "type": "string",
                  "description": "Quantity as string to support values like \"1/3\", \"2+1/3\"",
                  "example": "2+1/3"
                },
                "unit": {
                  "type": "string",
                  "description": "Unit of measure (unrestricted string)",
                  "example": "cups"
                }
              }
            },
            "example": [
              {
                "name": "All-purpose flour",
                "quantity": "2",
                "unit": "cups"
              },
              {
                "name": "Butter",
                "quantity": "1/2",
                "unit": "cup"
              }
            ]
          },
          "instructionsMarkdown": {
            "type": "string",
            "description": "Recipe instructions in markdown",
            "example": "1. Preheat oven to 350°F.\n2. Mix dry ingredients..."
          }
        },
        "required": [
          "ingredients",
          "instructionsMarkdown"
        ],
        "example": {
          "ingredients": [
            {
              "name": "All-purpose flour",
              "quantity": "2",
              "unit": "cups"
            }
          ],
          "instructionsMarkdown": "1. Preheat oven to 350°F.\n2. Mix dry ingredients."
        }
      },
      "recipe-version": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for this recipe version",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "recipeId": {
            "type": "string",
            "description": "Id of the recipe this version belongs to",
            "example": "a1b2c3d4-e89b-12d3-a456-426614174001"
          },
          "content": {
            "type": "object",
            "description": "Version body — structured ingredients and markdown instructions",
            "properties": {
              "ingredients": {
                "type": "array",
                "description": "List of ingredients; quantity and unit are strings (e.g. \"2+1/3\", \"cups\")",
                "items": {
                  "type": "object",
                  "required": [
                    "name",
                    "quantity",
                    "unit"
                  ],
                  "properties": {
                    "name": {
                      "type": "string",
                      "description": "Ingredient name",
                      "example": "All-purpose flour"
                    },
                    "quantity": {
                      "type": "string",
                      "description": "Quantity as string to support values like \"1/3\", \"2+1/3\"",
                      "example": "2+1/3"
                    },
                    "unit": {
                      "type": "string",
                      "description": "Unit of measure (unrestricted string)",
                      "example": "cups"
                    }
                  }
                },
                "example": [
                  {
                    "name": "All-purpose flour",
                    "quantity": "2",
                    "unit": "cups"
                  },
                  {
                    "name": "Butter",
                    "quantity": "1/2",
                    "unit": "cup"
                  }
                ]
              },
              "instructionsMarkdown": {
                "type": "string",
                "description": "Recipe instructions in markdown",
                "example": "1. Preheat oven to 350°F.\n2. Mix dry ingredients..."
              }
            },
            "required": [
              "ingredients",
              "instructionsMarkdown"
            ],
            "example": {
              "ingredients": [
                {
                  "name": "All-purpose flour",
                  "quantity": "2",
                  "unit": "cups"
                }
              ],
              "instructionsMarkdown": "1. Preheat oven to 350°F.\n2. Mix dry ingredients."
            }
          },
          "isLatest": {
            "type": "boolean",
            "description": "Whether this version is the current one for the recipe (only one per recipe is true)",
            "example": true
          },
          "createdBy": {
            "type": "string",
            "description": "User id of the creator of this version",
            "example": "a1b2c3d4-e89b-12d3-a456-426614174001"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "When this version was created",
            "example": "2023-01-15T14:30:00Z"
          }
        },
        "required": [
          "id",
          "recipeId",
          "content",
          "isLatest",
          "createdBy",
          "createdAt"
        ]
      },
...
      "unsplash-attribution": {
        "type": "object",
        "description": "Minimal fields needed to attribute an Unsplash photo: photographer and links with UTM params per Unsplash API requirements. Shown when displaying Unsplash-sourced recipe files (e.g. full-size view).",
        "properties": {
          "photographerName": {
            "type": "string",
            "description": "Display name of the photographer",
            "example": "Jane Doe"
          },
          "photographerProfileUrl": {
            "type": "string",
            "format": "uri",
            "description": "Unsplash profile URL with UTM params (utm_source, utm_medium=referral)",
            "example": "https://unsplash.com/@janedoe?utm_source=your_app_name&utm_medium=referral"
          },
          "unsplashSourceUrl": {
            "type": "string",
            "format": "uri",
            "description": "Unsplash source link with UTM params for \"on Unsplash\" attribution",
            "example": "https://unsplash.com?utm_source=your_app_name&utm_medium=referral"
          }
        },
        "required": [
          "photographerName",
          "photographerProfileUrl",
          "unsplashSourceUrl"
        ]
      },
      "recipe-file-info": {
        "type": "object",
        "description": "Metadata for one file attached to a recipe. Uses SAF file metadata (blob_name, file_original_name, mimetype, size, created_at, updated_at).",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the file record",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "recipeId": {
            "type": "string",
            "description": "Id of the recipe this file is attached to",
            "example": "a1b2c3d4-e89b-12d3-a456-426614174001"
          },
          "blobName": {
            "type": "string",
            "description": "Storage blob name (Azure) for the file",
            "example": "recipes/a1b2c3d4-e89b-12d3-a456-426614174001/123e4567-e89b-12d3-a456-426614174000.pdf"
          },
          "fileOriginalName": {
            "type": "string",
            "description": "Original filename as provided on upload",
            "example": "grandmas-cookies.pdf"
          },
          "mimetype": {
            "type": "string",
            "description": "MIME type of the file",
            "example": "application/pdf"
          },
          "size": {
            "type": "integer",
            "minimum": 0,
            "description": "File size in bytes",
            "example": 102400
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "When the file was uploaded",
            "example": "2023-01-15T14:30:00Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "When the file metadata was last updated",
            "example": "2023-01-15T14:30:00Z"
          },
          "uploadedBy": {
            "type": "string",
            "nullable": true,
            "description": "User id of the uploader (optional)",
            "example": "a1b2c3d4-e89b-12d3-a456-426614174001"
          },
          "downloadUrl": {
            "type": "string",
            "format": "uri",
            "description": "Full URL to download or view the file (API base + /recipes/:id/files/:fileId/blob)",
            "example": "https://api.recipes.example.com/recipes/a1b2c3d4-e89b-12d3-a456-426614174001/files/123e4567-e89b-12d3-a456-426614174000/blob"
          },
          "unsplashAttribution": {
            "$ref": "#/components/schemas/unsplash-attribution",
            "description": "When present, the file was sourced from Unsplash; frontend must show attribution (e.g. full-size view). Omitted when file is from upload."
          }
        },
        "required": [
          "id",
          "recipeId",
          "blobName",
          "fileOriginalName",
          "mimetype",
          "size",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "add-recipe-file-from-unsplash-request": {
        "type": "object",
        "description": "Payload to add a recipe file from an Unsplash photo. Backend will fetch the photo by ID to get the user object for storage.",
        "properties": {
          "unsplashPhotoId": {
            "type": "string",
            "description": "Unsplash photo id (short id) of the chosen photo",
            "example": "K3m9_xR2"
          },
          "downloadLocation": {
            "type": "string",
            "format": "uri",
            "description": "Unsplash download tracking URL; backend calls trackDownload with this",
            "example": "https://api.unsplash.com/photos/abc123/download"
          },
          "imageUrl": {
            "type": "string",
            "format": "uri",
            "description": "URL to fetch image bytes (e.g. regular or full size from search result)",
            "example": "https://images.unsplash.com/photo-123/regular"
          }
        },
        "required": [
          "unsplashPhotoId",
          "downloadLocation",
          "imageUrl"
        ]
      },
...
    }
  }
}